os/mac: move legacy functions to compat
