+++ Library/Homebrew/compat.rb
@@ -4,3 +4,4 @@ require "compat/cask/dsl/version"
 require "compat/language/python"
 require "compat/requirements/macos_requirement"
 require "compat/formula"
+require "compat/os/mac" if OS.mac?
